+++ R/full_geo_metadata.R
@@ -1,0 +1,15 @@
+#' Equivalencias de códigos de identificación geográfica
+#'
+#' Un archivo de datos que contiene identificadores geográficos para los departametnos de Argentina y sus provincias
+#' 
+#'
+#' @format Un tibble con 526 filas y 8 variables:
+#' \describe{
+#'   \item{codprov, coddepto}{identificadores utilizados por \emph{INDRA} para escrutinios provisorios de elecciones nacionales}
+#'   \item{*_censo}{identificadores de unidades geográficas del \emph{INDEC_}}
+#'   \item{*_iso}{identificadores de {International Organization for Standardization}}
+#'   
+#'   ...
+#' }
+#' 
+"full_geo_metadata"