+++ R/get_grid.R
@@ -1,6 +1,6 @@
 #'Carga grillas de districts de Argentina
 #'@description
-#'Función que descaga grillas (\emph{facet}) para acomodarlas como si fueran mapas de provincias y deparamentos
+#'Función que descarga grillas (\emph{facet}) para acomodarlas como si fueran mapas de provincias y deparamentos
 #' de Argentina.
 #' 
 #'