+++ NEWS.md
@@ -2,8 +2,13 @@
 
 **DRAFT, 2020: dev version**
 
-* (Already comitted and installed locally) Add 'level' parameter to 'get_multiple_election' function
+* (Already comitted and installed locally) Add 'level' parameter to `get_multiple_election()` function
+
 * Add new computation function for electoral data:  `compute_concentration()`
+
+* Fixed `map_results()`: need to ungroup when using `sf::st_as_sf()` function. 
+
+* Add `{geofaceteAR}` functions as `{polAr}` native functions. 
 
 **Jul 26, 2020: dev version**